rev. extract Container to abstract class
