+++ src/test/java/ru/petrov/repository/TypeOfValueRepositoryTest.java
@@ -1,18 +1,11 @@
 package ru.petrov.repository;
 
-import liquibase.exception.LiquibaseException;
-import liquibase.integration.spring.SpringLiquibase;
-import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 import ru.petrov.model.TypeOfValue;
 import ru.petrov.repository.jdbc.JdbcTypeOfValueRepository;
 
@@ -22,39 +15,16 @@ import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-@Testcontainers
-public class TypeOfValueRepositoryTest {
+public class TypeOfValueRepositoryTest extends AbstractRepositoryTest{
     public static TypeOfValueRepository typeOfValueRepository;
 
     TypeOfValue hotWater;
     TypeOfValue coldWater;
     TypeOfValue coldWaterNew;
-    @Container
-    private static final PostgreSQLContainer<?> postgresqlContainer = new PostgreSQLContainer<>("postgres:15.5");
 
     @BeforeAll
-    public static void startContainer() {
-        postgresqlContainer.start();
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName(postgresqlContainer.getDriverClassName());
-        dataSource.setUrl(postgresqlContainer.getJdbcUrl());
-        dataSource.setUsername(postgresqlContainer.getUsername());
-        dataSource.setPassword(postgresqlContainer.getPassword());
-        SpringLiquibase liquibase = new SpringLiquibase();
-        liquibase.setChangeLog("classpath:db/changelog/changelog-for-test.xml");
-        liquibase.setDataSource(dataSource);
-        try {
-            liquibase.afterPropertiesSet();
-            System.out.println("Changelog applied successfully.");
-        } catch (LiquibaseException e) {
-            System.err.println("Error applying changelog: " + e.getMessage());
-        }
+    public static void setRepository() {
         typeOfValueRepository = new JdbcTypeOfValueRepository(new JdbcTemplate(dataSource));
-    }
-
-    @AfterAll
-    public static void stopContainer() {
-        postgresqlContainer.stop();
     }
 
     @BeforeEach
